Move the quiz component
